requirements.txt
tabula-py
git+https://github.com/vscherbo/apps.git
